# wsfl-frontend/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "WSFL (Winter Series Fantasy League) API",
        "description": "API for managing fantasy running league competitions",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://choosealicense.com/licenses/mit/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/api/matchups": {
            "get": {
                "tags": [
                    "Matchup"
                ],
                "summary": "Get all matchups",
                "description": "Retrieves a paginated list of all matchups in the system",
                "operationId": "getAllMatchups",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "description": "Page number (0-based)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    },
                    {
                        "name": "raceId",
                        "in": "query",
                        "description": "Filter by race ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully retrieved matchups",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageMatchupSummaryDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Matchup"
                ],
                "summary": "Create a new matchup",
                "description": "Creates a new matchup in the system",
                "operationId": "createMatchup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchupCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Matchup created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchupDetailsDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid input data"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to create matchups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/matchups/{id}": {
            "get": {
                "tags": [
                    "Matchup"
                ],
                "summary": "Get a matchup by ID",
                "description": "Retrieves a specific matchup by its ID",
                "operationId": "getMatchup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the matchup",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Matchup not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Matchup not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Matchup found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchupDetailsDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Matchup"
                ],
                "summary": "Delete a matchup",
                "description": "Deletes a matchup from the system",
                "operationId": "deleteMatchup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the matchup",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Not authorized to delete matchups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Successfully deleted matchup"
                    },
                    "404": {
                        "description": "Matchup not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Matchup not found with id: 123"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Matchup"
                ],
                "summary": "Update matchup scores",
                "description": "Updates the scores for both teams in a matchup",
                "operationId": "updateMatchupScores",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the matchup",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchupPatchDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "403": {
                        "description": "Not authorized to update matchups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Matchup not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Matchup not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Scores updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchupDetailsDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TeamSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Team's unique identifier",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Team's name",
                        "example": "Thunderbolts"
                    },
                    "wins": {
                        "type": "integer",
                        "description": "Number of wins",
                        "format": "int32",
                        "example": 10
                    },
                    "losses": {
                        "type": "integer",
                        "description": "Number of losses",
                        "format": "int32",
                        "example": 5
                    },
                    "ties": {
                        "type": "integer",
                        "description": "Number of ties",
                        "format": "int32",
                        "example": 2
                    },
                    "totalScore": {
                        "type": "integer",
                        "description": "Total score",
                        "format": "int32",
                        "example": 157
                    }
                },
                "description": "Summary representation of a team"
            },
...
            "RaceSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Race's unique identifier",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the race",
                        "example": "Spring Championship"
                    },
                    "date": {
                        "type": "string",
                        "description": "Date of the race",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "isPlayoff": {
                        "type": "boolean",
                        "description": "Whether this is a playoff race",
                        "example": false
                    }
                },
                "description": "Summary representation of a race"
            },
...
            "MatchupCreateDto": {
                "required": [
                    "raceId",
                    "team1Id",
                    "team2Id"
                ],
                "type": "object",
                "properties": {
                    "raceId": {
                        "type": "integer",
                        "description": "ID of the race this matchup belongs to",
                        "format": "int64",
                        "example": 1
                    },
                    "team1Id": {
                        "type": "integer",
                        "description": "ID of the first team in the matchup",
                        "format": "int64",
                        "example": 1
                    },
                    "team2Id": {
                        "type": "integer",
                        "description": "ID of the second team in the matchup",
                        "format": "int64",
                        "example": 2
                    },
                    "team1Score": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Score of the first team",
                        "format": "int32",
                        "example": 100
                    },
                    "team2Score": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Score of the second team",
                        "format": "int32",
                        "example": 95
                    }
                },
                "description": "Matchup creation data"
            },
            "MatchupDetailsDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the matchup",
                        "format": "int64"
                    },
                    "race": {
                        "$ref": "#/components/schemas/RaceSummaryDto"
                    },
                    "team1": {
                        "$ref": "#/components/schemas/TeamSummaryDto"
                    },
                    "team2": {
                        "$ref": "#/components/schemas/TeamSummaryDto"
                    },
                    "team1Score": {
                        "type": "integer",
                        "description": "Score of the first team",
                        "format": "int32"
                    },
                    "team2Score": {
                        "type": "integer",
                        "description": "Score of the second team",
                        "format": "int32"
                    }
                },
                "description": "Detailed information about a matchup"
            },
...
            "MatchupPatchDto": {
                "type": "object",
                "properties": {
                    "team1Score": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Score of the first team",
                        "format": "int32",
                        "example": 100
                    },
                    "team2Score": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Score of the second team",
                        "format": "int32",
                        "example": 95
                    }
                },
                "description": "Matchup patch data"
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            },
...
            "MatchupSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the matchup",
                        "format": "int64"
                    },
                    "raceId": {
                        "type": "integer",
                        "description": "ID of the race this matchup belongs to",
                        "format": "int64"
                    },
                    "team1": {
                        "$ref": "#/components/schemas/TeamSummaryDto"
                    },
                    "team2": {
                        "$ref": "#/components/schemas/TeamSummaryDto"
                    },
                    "team1Score": {
                        "type": "integer",
                        "description": "Score of the first team",
                        "format": "int32"
                    },
                    "team2Score": {
                        "type": "integer",
                        "description": "Score of the second team",
                        "format": "int32"
                    }
                },
                "description": "Summary information about a matchup"
            },
            "PageMatchupSummaryDto": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MatchupSummaryDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}